+++ datavec-api/src/main/java/org/datavec/api/writable/batch/NDArrayRecordBatch.java
@@ -43,7 +43,7 @@ import java.util.List;
 public class NDArrayRecordBatch extends AbstractWritableRecordBatch {
 
     private List<INDArray> arrays;
-    private int size;
+    private long size;
 
     public NDArrayRecordBatch(INDArray... arrays){
         this(Arrays.asList(arrays));
@@ -68,13 +68,13 @@ public class NDArrayRecordBatch extends AbstractWritableRecordBatch {
 
     @Override
     public int size() {
-        return size;
+        return (int) size;
     }
 
     @Override
     public List<Writable> get(int index) {
         Preconditions.checkArgument(index >= 0 && index < size, "Invalid index: " + index + ", size = " + size);
-        List<Writable> out = new ArrayList<>(size);
+        List<Writable> out = new ArrayList<>((int) size);
         for (INDArray orig : arrays) {
             INDArray view = getExample(index, orig);
             out.add(new NDArrayWritable(view));
