+++ datavec-data/datavec-data-image/src/main/java/org/datavec/image/mnist/draw/DrawReconstruction.java
@@ -65,7 +65,7 @@ public class DrawReconstruction {
 
     public void draw() {
         WritableRaster r = img.getRaster();
-        int[] equiv = new int[data.length()];
+        int[] equiv = new int[(int) data.length()];
         INDArray dataLinear = data.linearView();
         for (int i = 0; i < equiv.length; i++)
             equiv[i] = Math.round(dataLinear.getInt(i));
